pyproject.toml
[project]
name = "django-ninja-crane"
version = "0.1.0"
description = "Add versioning and migrations for your Django Ninja API."
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-ninja>=1.5.1",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.8",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.11.1",
    "ruff>=0.14.10",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocs-callouts>=1.14",
    "plantuml-markdown>=3.10",
    "mike>=2.1.3",
    "mkdocs-awesome-nav>=3.3.0",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "test_server.settings"
asyncio_mode = "auto"

[tool.ty.src]
exclude = ["tests/", "test_app/", "test_server/"]
